Make getElementHeight a dom utility
